Projects/markdown2pdf/pyproject.toml
[project]
name = "markdown2pdf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.8",
    "markdown-pdf>=1.3.3",
    "markdown[standard]>=3.7",
    "requests>=2.32.3",
    "spire-doc>=13.1.0",
    "uvicorn[standard]>=0.34.0",
]
